Refactor and test of barrani deposit

+++ contracts/src/private-token/PublicKeyInfrastructure.sol
@@ -1,7 +1,8 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity 0.8.26;
 
-struct PublicKey { // #TODO : We could pack those in a single uint256 to save storage costs (for e.g using circomlibjs library to pack points on BabyJubjub) 
+struct PublicKey {
+    // #TODO : We could pack those in a single uint256 to save storage costs (for e.g using circomlibjs library to pack points on BabyJubjub)
     uint256 X;
     uint256 Y;
 } // The Public Key should be a point on Baby JubJub elliptic curve : checks must be done offchain before registering to ensure that X<p and Y<p and (X,Y) is on the curve
@@ -9,27 +10,26 @@ struct PublicKey { // #TODO : We could pack those in a single uint256 to save st
 
 /**
  * @dev Implementation of a simple PublicKeyInfrastructure.
- * The Public Key should be a point on the Baby JubJub elliptic curve : 
+ * The Public Key should be a point on the Baby JubJub elliptic curve :
  * It is represented by its coordinates X and Y in the struct PublicKey.
  * Any Ethereum account could set its own PublicKey, which is *non-revocable* once set.
  * *Important* : it is the responsability of the user to register a valid point on Baby Jubjub as his public key, or else transfers
  * to him will fail as there is no check on X and Y values to verify that the registered point is on Baby Jubjub in this contract.
  */
 contract PublicKeyInfrastructure {
-    
-    mapping(address  => PublicKey) internal registry;
+    mapping(address => PublicKey) internal registry;
 
-    event NewRegisteredPublicKey(address indexed owner, uint PubKeyX, uint PubKeyY);
+    event NewRegisteredPublicKey(address indexed owner, uint256 PubKeyX, uint256 PubKeyY);
 
     function registerPublicKey(uint256 X, uint256 Y) external {
-        require(X+Y!=0, "Public Key cannot be the origin point");
+        require(X + Y != 0, "Public Key cannot be the origin point");
         PublicKey memory registeredKey = registry[msg.sender];
-        require(registeredKey.X+registeredKey.Y==0, "Account already registered");
-        registry[msg.sender] = PublicKey(X,Y);
+        require(registeredKey.X + registeredKey.Y == 0, "Account already registered");
+        registry[msg.sender] = PublicKey(X, Y);
         emit NewRegisteredPublicKey(msg.sender, X, Y);
     }
 
-    function getRegistredKey(address user) external view returns(PublicKey memory){
+    function getRegistredKey(address user) external view returns (PublicKey memory) {
         return registry[user];
     }
 }
